hooked up view switching
